requirements.txt
fastapi
uvicorn
pytest
ruff
black
pytest-cov
pip-audit
pre-commit

# MLOps & Monitoring
mlflow  # MLflow for experiment tracking and model registry
evidently
prometheus-client

# Data Science & ML
pandas
numpy
requests
scikit-learn
xgboost  # Gradient boosting for ECG classification
matplotlib  # For plotting confusion matrices (optional but good)

# RAG & LLM (Phase 2)
langchain
langchain-community
langchain-huggingface
langchain-chroma  # ChromaDB integration (replaces deprecated langchain_community.vectorstores.Chroma)
chromadb  # Vector database for RAG
pypdf  # PDF loading
sentence-transformers  # Local embeddings (all-MiniLM-L6-v2)
python-dotenv  # Environment variable management
